separated the create package dialog into its own component
